Refactor the code related to spdm and pci_doe
